+++ src/app/mod.rs
@@ -1870,7 +1870,7 @@ impl App {
                     };
 
                     renderer.resize(&self.gpu.device, &mut resize_encoder, &self.gpu.queue, viewport_size.0, viewport_size.1,
-                        resize_source, self.config_manager.system_settings().iterations_per_thread, resize_config.zoom, resize_config.pan_x, resize_config.pan_y, resize_config.rotation,
+                        resize_source, self.config_manager.system_settings().iterations_per_thread, resize_config.zoom, resize_config.pan_x as f32, resize_config.pan_y as f32, resize_config.rotation,
                         resize_config.camera_rotation_x, resize_config.camera_rotation_y, resize_config.camera_bank, resize_config.camera_x, resize_config.camera_y, resize_config.camera_z, resize_config.speed_factor);
                     self.gpu.queue.submit(std::iter::once(resize_encoder.finish()));
 
@@ -1880,7 +1880,7 @@ impl App {
                     // Restore palette and color mode after buffer recreation
                     renderer.update_palette(&self.gpu.device, &self.gpu.queue, &resize_config.palette, resize_config.palette_rotation, resize_config.palette_squeeze, resize_config.palette_squeeze_mode, resize_config.palette_squeeze_falloff, resize_config.palette_log_strength, resize_config.palette_reverse);
                     renderer.set_color_mode(&self.gpu.queue, resize_config.color_mode, self.config_manager.system_settings().iterations_per_thread, self.config_manager.system_settings().burn_in,
-                        resize_config.zoom, resize_config.pan_x, resize_config.pan_y, resize_config.rotation,
+                        resize_config.zoom, resize_config.pan_x as f32, resize_config.pan_y as f32, resize_config.rotation,
                         resize_config.camera_rotation_x, resize_config.camera_rotation_y, resize_config.camera_bank, resize_config.camera_x, resize_config.camera_y, resize_config.camera_z, resize_config.speed_factor);
                     renderer.set_path_map_style(resize_config.path_map_style);
 
@@ -2204,8 +2204,8 @@ impl App {
                             iterations_per_thread,
                             20, // burn_in - use default for WASM export
                             export_config.zoom,
-                            export_config.pan_x,
-                            export_config.pan_y,
+                            export_config.pan_x as f32,
+                            export_config.pan_y as f32,
                             export_config.rotation,
                             export_config.camera_rotation_x,
                             export_config.camera_rotation_y,
@@ -3040,8 +3040,8 @@ impl App {
             if should_iterate
                 && renderer.maybe_refit_shadow(
                     final_config.zoom,
-                    final_config.pan_x,
-                    final_config.pan_y,
+                    final_config.pan_x as f32,
+                    final_config.pan_y as f32,
                     final_config.camera_rotation_x,
                     final_config.camera_rotation_y,
                     final_config.camera_bank,
@@ -3053,8 +3053,8 @@ impl App {
                     &self.gpu.queue,
                     self.config_manager.system_settings().iterations_per_thread,
                     final_config.zoom,
-                    final_config.pan_x,
-                    final_config.pan_y,
+                    final_config.pan_x as f32,
+                    final_config.pan_y as f32,
                     final_config.rotation,
                     final_config.camera_rotation_x,
                     final_config.camera_rotation_y,
@@ -3191,7 +3191,7 @@ impl App {
 
                 let samples_this_frame = renderer.compute_pass(&mut render_encoder, &self.gpu.queue, &self.gpu.device, effective_workgroups,
                     self.config_manager.system_settings().iterations_per_thread, self.config_manager.system_settings().burn_in,
-                    final_config.zoom, final_config.pan_x, final_config.pan_y, final_config.rotation,
+                    final_config.zoom, final_config.pan_x as f32, final_config.pan_y as f32, final_config.rotation,
                     final_config.camera_rotation_x, final_config.camera_rotation_y, final_config.camera_bank, final_config.camera_x, final_config.camera_y, final_config.camera_z, final_config.speed_factor, clear_histogram, clear_paths);
 
                 self.metrics.record_compute_time(t_compute.elapsed().as_secs_f64() * 1000.0);
@@ -3273,8 +3273,8 @@ impl App {
                 &mut render_encoder,
                 final_config.zoom,
                 final_config.rotation,
-                final_config.pan_x,
-                final_config.pan_y,
+                final_config.pan_x as f32,
+                final_config.pan_y as f32,
                 final_config.camera_rotation_x,
                 final_config.camera_rotation_y,
                 final_config.camera_bank,
@@ -3599,7 +3599,10 @@ impl App {
         width: u32,
         height: u32,
         config: &crate::config::FractalConfig,
-    ) -> (f32, f32) {
+        // f64 out, because this is an absolute position in the
+        // fractal plane and that is the one thing a deep zoom
+        // destroys -- the same reason `pan_x` is f64.
+    ) -> (f64, f64) {
         // Convert pixel to normalized device coordinates (-1 to 1)
         let ndc_x = (pixel_x as f32 / width as f32) * 2.0 - 1.0;
         let ndc_y = (pixel_y as f32 / height as f32) * 2.0 - 1.0;
@@ -3610,11 +3613,11 @@ impl App {
         let scaled_y = ndc_y;
 
         // Screen space → pan frame (rotation-aware in 2D, identity in 3D)
-        let (rotated_x, rotated_y) = config.screen_delta_to_pan_frame(scaled_x, scaled_y);
+        let (rotated_x, rotated_y) = config.screen_delta_to_pan_frame(scaled_x as f64, scaled_y as f64);
 
         // Apply inverse zoom and add pan
-        let fractal_x = rotated_x / config.zoom + config.pan_x;
-        let fractal_y = rotated_y / config.zoom + config.pan_y;
+        let fractal_x = rotated_x / config.zoom as f64 + config.pan_x;
+        let fractal_y = rotated_y / config.zoom as f64 + config.pan_y;
 
         (fractal_x, fractal_y)
     }
